+++ tinboard/app.py
@@ -16,7 +16,8 @@ from textual.binding import Binding
 ##############################################################################
 # Local imports.
 from .data import ExitStates, load_configuration, save_configuration, token_file
-from .screens import Main, TokenInput
+from .pinboard import API, BookmarkData
+from .screens import BookmarkInput, Main, TokenInput
 from .widgets.filters import Filters
 
 
@@ -52,7 +53,7 @@ class Tinboard(App[ExitStates]):
         """
         if token:
             token_file().write_text(token, encoding="utf-8")
-            self.push_screen(Main(token))
+            self.push_screen(Main(API(token), self._initial_filter))
         else:
             self.exit(ExitStates.TOKEN_NEEDED)
 
@@ -81,6 +82,20 @@ class Tinboard(App[ExitStates]):
             pass
         return None
 
+    async def inline_add(self, result: BookmarkData | None = None) -> None:
+        """Handle the result adding a bookmark inline.
+
+        Args:
+            result: The result data, or `None` if the entry was cancelled.
+        """
+        if result:
+            try:
+                if token := self.api_token:
+                    await API(token).add_bookmark(result)
+            except API.Error:
+                self.exit(ExitStates.INLINE_SAVE_ERROR)
+        self.exit()
+
     def on_mount(self) -> None:
         """Display the main screen.
 
@@ -90,7 +105,17 @@ class Tinboard(App[ExitStates]):
             once the token has been acquired.
         """
         if token := self.api_token:
-            self.push_screen(Main(token, self._initial_filter))
+            if self.is_inline:
+                # If we're running as an inline application, that means we
+                # should simply be showing the bookmark input screen for
+                # quick input. Note disabling the command palette too, which
+                # has problems with inline mode.
+                self.use_command_palette = False
+                self.push_screen(BookmarkInput(API(token)), callback=self.inline_add)
+            else:
+                # We're not in inline mode so we'll show the normal
+                # application screen.
+                self.push_screen(Main(API(token), self._initial_filter))
         else:
             self.push_screen(TokenInput(), callback=self.token_bounce)